+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## unreleased
+## 0.15.0 - 2024-10-11
 ### Added
 * Added support for toon gradient textures to generated materials.
 * Added support for editing the selected toon gradient ramp for wimdo exports.